Resolve bug
